face_attendance/requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
asgiref==3.10.0
astunparse==1.6.3
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
cython==3.1.4
deepface==0.0.95
django==4.2.25
djangorestframework==3.16.1
# dlib @ file:///C:/Users/Hammad/Desktop/face_attendance/Dlib-main/Dlib-main/dlib-19.22.99-cp39-cp39-win_amd64.whl
easydict==1.13
eval-type-backport==0.2.2
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.19.1
fire==0.7.1
flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.75.1
gunicorn==23.0.0
h5py==3.14.0
humanfriendly==10.0
idna==3.11
imageio==2.37.0
importlib-metadata==8.7.0
importlib-resources==6.5.2
insightface==0.7.3
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.2
keras==3.10.0
kiwisolver==1.4.7
lazy-loader==0.4
libclang==18.1.1
markdown==3.9
markdown-it-py==3.0.0
markupsafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.5.3
mpmath==1.3.0
mtcnn==0.1.1
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
onnx==1.19.1
onnxruntime==1.19.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt-einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
prettytable==3.16.0
protobuf==6.33.0
pydantic==2.12.3
pydantic-core==2.41.4
pygments==2.19.2
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
simsimd==6.5.3
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
stringzilla==4.2.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf-keras==2.20.1
threadpoolctl==3.6.0
tifffile==2024.8.30
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
